requirements.txt
nest_asyncio
duckduckgo-search
transformers
phidata
newspaper4k
lxml_html_clean
duckduckgo_search
transformers
datasets
pandas
nltk
torch
tqdm
tiktoken 
sentencepiece
gensim
soundfile
gtts
deep_translator
fastapi
pydantic
uvicorn
streamlit